% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{git_pre_commit}
\alias{git_pre_commit}
\alias{add_pre_commit}
\alias{add_pre_push}
\title{Adds a git hook}
\usage{
git_pre_commit()

add_pre_commit()

add_pre_push()
}
\description{
git_pre_commit only tests files that have changed.
Makes things, especially linting a lot faster.

Add a pre-commit or pre-push hook.
Both hooks are identical. They just get triggered at different times.
}
