% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rproj.R
\name{get_default_rproj}
\alias{get_default_rproj}
\alias{check_rproj}
\title{Check the .Rproj file}
\usage{
get_default_rproj()

check_rproj(default_rproj = get_default_rproj(), path = ".")
}
\arguments{
\item{default_rproj}{Default Rproj values.}

\item{path}{Location of the package}
}
\description{
If a .Rproj file exists, then the values should correspond to those given by
\code{get_default_rproj}. If more than one .Rproj file exists, an error is raised.
}
\examples{
check_readme()
}
