% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_renviron_var.R
\name{set_renviron_var}
\alias{set_renviron_var}
\title{Sets an Renviron variable}
\usage{
set_renviron_var(variable, value)
}
\arguments{
\item{variable}{The variable name}

\item{value}{The value}
}
\description{
Set an variable in the .Renviron file. Calls \code{readEnviron} to enable R
to use the variable in future function calls.
}
